Editor: Toolboxes work - Added ability to enforce default visibility state on each editor startup by config pack - Rename "* item toolbox" into "* Item browser" - New icons for item browser toolboxes - New theme icons to override item browsers icons

+++ Editor/main_window/tools/app_settings.cpp
@@ -207,7 +207,7 @@ void g_AppSettings::on_buttonBox_accepted()
     GlobalSettings::Placing_dontShowPropertiesBox = ui->Editor_placing_dontShowPropsBox->isChecked();
     GlobalSettings::historyLimit = ui->historyLimit->value();
 
-    GlobalSettings::LvlOpts.default_zoom = ui->defaultZoom->value();
+    GlobalSettings::LvlOpts.default_zoom = static_cast<unsigned int>(ui->defaultZoom->value());
 
     if(ui->screengrabFit->isChecked())
         GlobalSettings::screenGrab.sizeType = SETTINGS_ScreenGrabSettings::GRAB_Fit;
@@ -254,7 +254,7 @@ void g_AppSettings::on_buttonBox_accepted()
         GlobalSettings::LvlItemDefaults.npc_direction=-1;
 
     GlobalSettings::LvlItemDefaults.npc_generator_type = ui->defaults_npc_generator_type->currentIndex()+1;
-    GlobalSettings::LvlItemDefaults.npc_generator_delay = round(ui->defaults_npc_generator_delay->value()*10.0);
+    GlobalSettings::LvlItemDefaults.npc_generator_delay = static_cast<int>(round(ui->defaults_npc_generator_delay->value() * 10.0));
     GlobalSettings::LvlItemDefaults.warp_type = ui->defaults_warps_type->currentIndex();
 
     GlobalSettings::LvlItemDefaults.classicevents_tabs_layviz=ui->defaults_eventtab_layviz->isChecked();
